+++ features/location/impl/src/main/kotlin/io/element/android/features/location/impl/common/permissions/PermissionsPresenter.kt
@@ -14,8 +14,12 @@
  * limitations under the License.
  */
 
-package io.element.android.features.location.impl.permissions
+package io.element.android.features.location.impl.common.permissions
 
-sealed interface PermissionsEvents {
-    object RequestPermissions : PermissionsEvents
+import io.element.android.libraries.architecture.Presenter
+
+interface PermissionsPresenter : Presenter<PermissionsState> {
+    interface Factory {
+        fun create(permissions: List<String>): PermissionsPresenter
+    }
 }
