+++ features/location/impl/src/test/kotlin/io/element/android/features/location/impl/common/actions/AndroidLocationActionsTest.kt
@@ -14,11 +14,11 @@
  * limitations under the License.
  */
 
-package io.element.android.features.location.impl.show
+package io.element.android.features.location.impl.common.actions
 
 import com.google.common.truth.Truth.assertThat
 import io.element.android.features.location.api.Location
-import io.element.android.features.location.impl.buildUrl
+import io.element.android.features.location.impl.common.actions.buildUrl
 import org.junit.Test
 import java.net.URLEncoder
